switch all dice rolls into the roll dice controller
