+++ WarhammerGameManager.Frontend/Controllers/GameTrackerController.cs
@@ -49,48 +49,5 @@ namespace WarhammerGameManager.Frontend.Controllers
 
             return RedirectToAction("OpenGame", new { gameId = newGameId });
         }
-
-        [HttpPost]
-        public async Task<IActionResult> RollDice(RollDiceRequest request, long? GameId = null)
-        {
-            DiceEvent data;
-
-            //If there is no provided GameId, then this is a quick roll
-            if (GameId == null)
-            {
-                data = await _gml.QuickRollDice(request);
-            }
-            else
-            {
-                data = await _gml.GameRoll(request, GameId.Value);
-                await _hubContext.Clients.All.SendAsync("ReceiveRollsUpdates", GameId.Value);
-            }
-
-            var response = new RollDiceResponse()
-            {
-                Request = request,
-                Results = data
-            };
-
-            return PartialView("~/Views/GameTracker/Partials/RollDiceResultsPartialView.cshtml", response);
-        }
-
-        [HttpPost]
-        public async Task<IActionResult> RollDiceBasic(RollDiceBasicRequest request, long? GameId = null)
-        {
-            RollDiceBasicResponse response;
-            //If there is no provided GameId, then this is a quick roll
-            if (GameId == null)
-            {
-                response = await _gml.QuickRollDice(request);
-            }
-            else
-            {
-                response = await _gml.GameRoll(request, GameId.Value);
-                await _hubContext.Clients.All.SendAsync("ReceiveRollsUpdates", GameId.Value);
-            }
-
-            return PartialView("~/Views/GameTracker/Partials/RollDiceResultsBasicPartialView.cshtml", response);
-        }
     }
 }
