+++ WarhammerGameManager.Frontend/Views/RollDice/Partials/RollDiceFormsPartial.cshtml
@@ -1,8 +1,8 @@
 ﻿@model List<RollType>
 
-<form id="RollDiceRequestForm" asp-controller="GameTracker" asp-action="RollDice" method="post">
+<form id="RollDiceRequestForm" asp-controller="RollDice" asp-action="RollDice" method="post">
 </form>
-<form id="RollDiceBasicRequestForm" asp-controller="GameTracker" asp-action="RollDiceBasic" method="post">
+<form id="RollDiceBasicRequestForm" asp-controller="RollDice" asp-action="RollDiceBasic" method="post">
 </form>
 
 <div class="row justify-content-center">
